requirements.txt
requests ~= 2.28.1
lxml ~= 4.9.2
python-dotenv ~= 0.21.1